+++ packages/breadcrumbs/src/BreadcrumbItem.ts
@@ -22,6 +22,7 @@ import { Focusable } from '@spectrum-web-components/shared/src/focusable.js';
 import { LikeAnchor } from '@spectrum-web-components/shared/src/like-anchor.js';
 import chevronStyles from '@spectrum-web-components/icon/src/spectrum-icon-chevron.css.js';
 import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron100.js';
+import chevronIconOverrides from '@spectrum-web-components/icon/src/icon-chevron-overrides.css.js';
 
 import styles from './breadcrumb-item.css.js';
 
@@ -31,7 +32,7 @@ export interface BreadcrumbSelectDetail {
 
 export class BreadcrumbItem extends LikeAnchor(Focusable) {
     public static override get styles(): CSSResultArray {
-        return [styles, chevronStyles];
+        return [styles, chevronStyles, chevronIconOverrides];
     }
 
     @property()
@@ -82,6 +83,12 @@ export class BreadcrumbItem extends LikeAnchor(Focusable) {
         }
     }
 
+    protected handleKeyDown(event: KeyboardEvent): void {
+        if (event.key === 'Enter' || event.keyCode === 13) {
+            this.handleClick(event);
+        }
+    }
+
     protected renderLink(): TemplateResult {
         return html`
             <a
@@ -91,6 +98,7 @@ export class BreadcrumbItem extends LikeAnchor(Focusable) {
                 aria-current=${ifDefined(
                     this.isLastOfType ? 'page' : undefined
                 )}
+                @keydown=${this.handleKeyDown}
                 @click=${this.handleClick}
             >
                 <slot></slot>