+++ packages/close-button/src/spectrum-config.js
@@ -60,15 +60,15 @@ const config = {
                         ['spectrum-CloseButton--staticWhite', 'white'],
                         ['spectrum-CloseButton--staticBlack', 'black'],
                     ],
-                    'static'
+                    'static-color'
                 ),
                 // Default to `size='m'` without needing the attribute
-                converter.classToHost('spectrum-Closebutton--sizeM'),
+                converter.classToHost('spectrum-CloseButton--sizeM'),
                 ...converter.enumerateAttributes(
                     [
-                        ['spectrum-Closebutton--sizeS', 's'],
-                        ['spectrum-Closebutton--sizeL', 'l'],
-                        ['spectrum-Closebutton--sizeXL', 'xl'],
+                        ['spectrum-CloseButton--sizeS', 's'],
+                        ['spectrum-CloseButton--sizeL', 'l'],
+                        ['spectrum-CloseButton--sizeXL', 'xl'],
                     ],
                     'size'
                 ),