+++ packages/popover/test/benchmark/test-basic.ts
@@ -11,16 +11,17 @@ governing permissions and limitations under the License.
 */
 
 import '@spectrum-web-components/popover/sp-popover.js';
+import '@spectrum-web-components/dialog/sp-dialog.js';
 import { html } from 'lit';
 import { measureFixtureCreation } from '../../../../test/benchmark/helpers.js';
 
 measureFixtureCreation(html`
-    <sp-popover variant="dialog" direction="top" open>
-        <div id="title">Popover Title</div>
-        <div id="content">
+    <sp-popover direction="top" open>
+        <sp-dialog>
+            <h3 slot="heading">Popover title</h3>
             Cupcake ipsum dolor sit amet jelly beans. Chocolate jelly caramels.
             Icing soufflé chupa chups donut cheesecake. Jelly-o chocolate cake
             sweet roll cake danish candy biscuit halvah
-        </div>
+        </sp-dialog>
     </sp-popover>
 `);